Add method recalculate_interest to scheduler.rake
Rename method simulation to trade_simulation
Move trade_simulation to /app/lib/trade_simulation.rb
