+++ config/application.rb
@@ -43,8 +43,7 @@ module TradeSimulator
                        request_specs: false
       g.fixture_replacement :factory_bot, dir: "spec/factories"
     end
-
+    
     config.time_zone = 'Tokyo'
-
   end
 end
