+++ db/seeds.rb
@@ -115,4 +115,5 @@ user_name.each_with_index do |name, i|
     end
 end
 end
+
 puts "create sample #{benchmark_result}s"
