--- spec/controllers/conditions_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe ConditionsController, type: :controller do
-
-end
