--- spec/controllers/favorites_controller_spec.rb
@@ -1,19 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe FavoritesController, type: :controller do
-
-  describe "GET #create" do
-    it "returns http success" do
-      get :create
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-  describe "GET #destroy" do
-    it "returns http success" do
-      get :destroy
-      expect(response).to have_http_status(:success)
-    end
-  end
-
-end
